+++ tests/test_lib.py
@@ -7,7 +7,6 @@ import os
 from collections import Counter, defaultdict, namedtuple
 from functools import partial
 from itertools import groupby, starmap
-from operator import truth
 
 import pytest
 from beetsplug.bandcamp import BandcampPlugin
@@ -27,7 +26,7 @@ pytestmark = pytest.mark.lib
 
 BASE_DIR = "lib_tests"
 TEST_DIR = "dev"
-REFERENCE_DIR = "97295a4"
+REFERENCE_DIR = "v0.15.1"
 JSONS_DIR = "jsons"
 
 IGNORE_FIELDS = {
@@ -42,6 +41,7 @@ IGNORE_FIELDS = {
     "artwork",
     "city",
     "disctitle",
+    "times_bought",
 }
 
 target_dir = os.path.join(BASE_DIR, TEST_DIR)
@@ -61,7 +61,7 @@ TRACK_FIELDS = ["track_alt", "artist", "title"]
 
 def album_table(**kwargs):
     table = new_table(*TRACK_FIELDS, show_header=False, highlight=False)
-    return border_panel(table, **{**dict(expand=True), **kwargs})
+    return border_panel(table, **{"expand": True, **kwargs})
 
 
 albums = defaultdict(album_table)
@@ -217,9 +217,17 @@ def test_file(file, guru, cache):
                 new = album
                 break
 
-    new.catalognum = " / ".join(filter(truth, map(lambda x: x.catalognum, guru.albums)))
-    with open(target_file, "w") as f:
-        json.dump(new, f, indent=2)
+    new.catalognum = " / ".join(x.catalognum for x in guru.albums if x.catalognum)
+    try:
+        with open(target_file) as f:
+            contents = json.load(f)
+    except FileNotFoundError:
+        with open(target_file, "w") as f:
+            json.dump(new, f, indent=2)
+    else:
+        if new != contents:
+            with open(target_file, "w") as f:
+                json.dump(new, f, indent=2)
 
     try:
         with open(os.path.join(compare_against, file)) as f: